+++ helpful.el
@@ -1234,6 +1234,25 @@ If the source code cannot be found, return the sexp used."
     (or (assoc sym completions)
         (assoc-string sym completions))))
 
+(defun helpful--version-info (sym)
+  "If SYM has version information, format and return it.
+Return nil otherwise."
+  (when (symbolp sym)
+    (let ((package-version
+           (get sym 'custom-package-version))
+          (emacs-version
+           (get sym 'custom-version)))
+      (cond
+       (package-version
+        (format
+         "This variable was added, or its default value changed, in %s version %s."
+         (car package-version)
+         (cdr package-version)))
+       (emacs-version
+        (format
+         "This variable was added, or its default value changed, in Emacs %s."
+         emacs-version))))))
+
 (defun helpful--library-path (library-name)
   "Find the absolute path for the source of LIBRARY-NAME.
 
@@ -2154,14 +2173,20 @@ state of the current symbol."
         (when (custom-variable-p helpful--sym)
           (insert " " (helpful--make-customize-button helpful--sym)))))
 
-    (-when-let (docstring (helpful--docstring helpful--sym helpful--callable-p))
-      (helpful--insert-section-break)
-      (insert
-       (helpful--heading "Documentation")
-       (helpful--format-docstring docstring))
-      (when (helpful--in-manual-p helpful--sym)
-        (insert "\n\n")
-        (insert (helpful--make-manual-button helpful--sym))))
+    (let ((docstring (helpful--docstring helpful--sym helpful--callable-p))
+          (version-info (unless helpful--callable-p
+                          (helpful--version-info helpful--sym))))
+      (when (or docstring version-info)
+        (helpful--insert-section-break)
+        (insert
+         (helpful--heading "Documentation"))
+        (when docstring
+          (insert (helpful--format-docstring docstring)))
+        (when version-info
+          (insert "\n\n" (s-word-wrap 70 version-info)))
+        (when (helpful--in-manual-p helpful--sym)
+          (insert "\n\n")
+          (insert (helpful--make-manual-button helpful--sym)))))
 
     ;; Show keybindings.
     ;; TODO: allow users to conveniently add and remove keybindings.