+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 Show the original value for custom variables whose value has changed.
 
+Report the package version when custom variables were added.
+
 # v0.17
 
 Fixed a minor docstring formatting issue when keymap references were