+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Show the original value for custom variables whose value has changed.
 
 Report the package version when custom variables were added.
 
+Fixed a crash on assigning byte-compiled objects to keybindings.
+
 # v0.17
 
 Fixed a minor docstring formatting issue when keymap references were