+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Fixed a crash on assigning byte-compiled objects to keybindings.
 
 Fixed an issue with advice being shown in docstrings on Emacs 27+.
 
+Symbol links in docstrings are now smarter in cases like "function `foo'".
+
 # v0.17
 
 Fixed a minor docstring formatting issue when keymap references were